+++ ErsatzTV.Application/MediaCollections/Mapper.cs
@@ -54,10 +54,11 @@ internal static class Mapper
             playlistGroups.Map(bg => new TreeGroupViewModel(
                 bg.Id,
                 bg.Name,
-                bg.Playlists.Map(b => new TreeItemViewModel(b.Id, b.Name)).ToList())).ToList());
+                bg.Playlists.Map(b => new TreeItemViewModel(b.Id, b.Name, b.IsSystem)).ToList(),
+                bg.IsSystem)).ToList());
 
     internal static PlaylistGroupViewModel ProjectToViewModel(PlaylistGroup playlistGroup) =>
-        new(playlistGroup.Id, playlistGroup.Name, playlistGroup.Playlists.Count);
+        new(playlistGroup.Id, playlistGroup.Name, playlistGroup.Playlists.Count, playlistGroup.IsSystem);
 
     internal static PlaylistViewModel ProjectToViewModel(Playlist playlist) =>
         new(playlist.Id, playlist.PlaylistGroupId, playlist.Name, playlist.IsSystem);